feat(core): add media projection audio source
