+++ core/src/main/java/io/github/thibaultbee/streampack/core/internal/sources/IMediaProjectionSource.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2024 Thibault B.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.github.thibaultbee.streampack.core.internal.sources
+
+import androidx.activity.result.ActivityResult
+
+/**
+ * Interface to implement class that uses MediaProjection.
+ * This interface is used to get the activity result from the activity.
+ */
+interface IMediaProjectionSource {
+    /**
+     * Set the activity result to get the media projection.
+     */
+    var activityResult: ActivityResult?
+}
